refactor(walk/dropper): one-file-per-concern + PredicateDescriptor trait
Split monolithic dropper.rs (1332 lines) into dropper/ module:
  gap.rs, template.rs, predicate.rs, emit.rs, verify.rs, error.rs,
  predicates/not_null.rs

Key design change: PredicateDescriptor trait encapsulates all
predicate-specific knowledge (detection, rendering, guard checking,
premise-guard detection). Adding a new predicate is a single new file
in predicates/; no changes to the pipeline.

Also fixes #405: detect_gaps now calls descriptor.is_premise_guarded()
and skips arrivals where the predicate is already conditionally
discharged via an Implies premise (e.g. callsite under
`if x.is_some() { f(x); }`). Three new tests cover guarded, unguarded,
and the conservative both-operands case.

Closes #405
