+++ implementations/rust/provekit-walk/src/dropper/predicates/mod.rs
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod not_null;
+pub use not_null::NotNullPredicate;
