+++ implementations/rust/provekit-walk/src/dropper/predicates/not_null.rs
@@ -1,0 +1,269 @@
+// SPDX-License-Identifier: Apache-2.0
+
+use provekit_ir_types::{IrFormula, IrTerm};
+
+use crate::dropper::predicate::{
+    formula_contains_predicate, predicate_var_arg, PredicateDescriptor,
+};
+use crate::dropper::template::{DropTemplate, NotRenderable};
+
+/// NotNullPredicate implements all predicate-specific knowledge for `not_null`.
+#[derive(Debug)]
+pub struct NotNullPredicate;
+
+impl PredicateDescriptor for NotNullPredicate {
+    fn name(&self) -> &str {
+        "not_null"
+    }
+
+    fn contains(&self, formula: &IrFormula) -> bool {
+        formula_contains_predicate(formula, "not_null")
+    }
+
+    fn var_arg(&self, formula: &IrFormula) -> Option<String> {
+        predicate_var_arg(formula, "not_null")
+    }
+
+    /// THE #405 FIX: returns true if `entry_wp` is
+    /// `Implies([premise, conclusion])` where:
+    /// - `conclusion` contains `not_null(var_name)`
+    /// - `premise` contains a guard for `var_name` (`is_some(var)` or
+    ///   `Not([is_none(var)])`)
+    /// - CONSERVATIVE: if `not_null` appears in BOTH premise AND conclusion,
+    ///   return false (emit the gap).
+    fn is_premise_guarded(&self, entry_wp: &IrFormula, var_name: &str) -> bool {
+        let IrFormula::Implies { operands } = entry_wp else {
+            return false;
+        };
+        if operands.len() < 2 {
+            return false;
+        }
+        let conclusion = &operands[operands.len() - 1];
+        let premises = &operands[..operands.len() - 1];
+
+        // Conclusion must contain not_null(var_name).
+        if !formula_contains_predicate(conclusion, "not_null") {
+            return false;
+        }
+        if predicate_var_arg(conclusion, "not_null").as_deref() != Some(var_name) {
+            return false;
+        }
+
+        // Conservative: if not_null appears in any premise, emit the gap.
+        if premises.iter().any(|p| formula_contains_predicate(p, "not_null")) {
+            return false;
+        }
+
+        // At least one premise must be a guard (is_some(var) or Not(is_none(var))).
+        premises.iter().any(|p| is_guard_for(p, var_name))
+    }
+
+    fn verified_templates(&self) -> &[DropTemplate] {
+        &[DropTemplate::Defensive]
+    }
+
+    fn render(&self, template: DropTemplate, var: &str) -> Result<String, NotRenderable> {
+        match template {
+            DropTemplate::Defensive => Ok(format!(
+                "    if {var}.is_none() {{ panic!(\"not_null: {var} must be Some\"); }}\n",
+                var = var
+            )),
+            DropTemplate::Recoverable => Err(NotRenderable::Scaffolding {
+                family: "Recoverable",
+                reason:
+                    "render emits `return Err(NullInput)` but no error type is defined; \
+                     pending caller-supplied error_expr support",
+            }),
+            DropTemplate::EarlyReturn => Err(NotRenderable::Scaffolding {
+                family: "EarlyReturn",
+                reason:
+                    "render emits `return Default::default()` which requires the caller's \
+                     return type to implement Default; not closure-verified by current lifter",
+            }),
+            DropTemplate::Expect => Err(NotRenderable::Scaffolding {
+                family: "Expect",
+                reason:
+                    "render shadows the original variable with a different type, breaking \
+                     downstream callsites; pending fresh-name binding (see issue #407)",
+            }),
+        }
+    }
+
+    fn guard_discharged(&self, formula: &IrFormula, var_name: &str) -> bool {
+        formula_contains_guard_for(formula, var_name)
+    }
+}
+
+/// Returns true if `formula` is a structural guard that discharges `not_null`
+/// for `var_name`:
+/// - `is_some(var_name)` (Atomic)
+/// - `Not([is_none(var_name)])` or `Not([is_some(var_name)])` (lift shape)
+fn is_guard_for(formula: &IrFormula, var_name: &str) -> bool {
+    match formula {
+        IrFormula::Atomic { name, args } => {
+            matches!(name.as_str(), "is_some") && has_var(args, var_name)
+        }
+        IrFormula::Not { operands } => {
+            if operands.len() == 1 {
+                if let IrFormula::Atomic { name, args } = &operands[0] {
+                    return matches!(name.as_str(), "is_none" | "is_some")
+                        && has_var(args, var_name);
+                }
+            }
+            false
+        }
+        _ => false,
+    }
+}
+
+fn has_var(args: &[IrTerm], var_name: &str) -> bool {
+    args.iter().any(|t| match t {
+        IrTerm::Var { name } => name == var_name,
+        _ => false,
+    })
+}
+
+/// Returns true if `formula` contains a `Not` node whose single operand is
+/// `Atomic { name: "is_none" | "is_some", args }` with a `Var` argument
+/// matching `var_name`. This is the exact shape lift.rs produces for
+/// if-then-panic.
+fn formula_contains_guard_for(formula: &IrFormula, var_name: &str) -> bool {
+    match formula {
+        IrFormula::Not { operands } => {
+            if operands.len() == 1 {
+                if let IrFormula::Atomic { name, args } = &operands[0] {
+                    let is_guard = matches!(name.as_str(), "is_none" | "is_some");
+                    let has_var = args.iter().any(|t| match t {
+                        IrTerm::Var { name } => name == var_name,
+                        _ => false,
+                    });
+                    if is_guard && has_var {
+                        return true;
+                    }
+                }
+            }
+            operands.iter().any(|o| formula_contains_guard_for(o, var_name))
+        }
+        IrFormula::Atomic { .. } => false,
+        IrFormula::And { operands }
+        | IrFormula::Or { operands }
+        | IrFormula::Implies { operands } => {
+            operands.iter().any(|o| formula_contains_guard_for(o, var_name))
+        }
+        IrFormula::Forall { body, .. } | IrFormula::Exists { body, .. } => {
+            formula_contains_guard_for(body, var_name)
+        }
+        IrFormula::Choice { body, .. } => formula_contains_guard_for(body, var_name),
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use provekit_ir_types::{IrFormula, IrTerm};
+
+    #[test]
+    fn not_null_predicate_verified_templates_returns_defensive_only() {
+        let templates = NotNullPredicate.verified_templates();
+        assert_eq!(templates.len(), 1, "one verified template for not_null (MVP)");
+        assert!(templates.contains(&DropTemplate::Defensive));
+    }
+
+    #[test]
+    fn defensive_template_renders_panic_shape() {
+        let rendered = NotNullPredicate
+            .render(DropTemplate::Defensive, "x")
+            .expect("Defensive must render OK");
+        assert!(rendered.contains("x.is_none()"), "must guard x");
+        assert!(rendered.contains("panic!"), "must panic on violation");
+        assert!(rendered.contains("not_null"), "panic msg must name invariant");
+    }
+
+    #[test]
+    fn recoverable_template_returns_not_renderable() {
+        let result = NotNullPredicate.render(DropTemplate::Recoverable, "x");
+        let err = result.expect_err("Recoverable must return NotRenderable");
+        match err {
+            NotRenderable::Scaffolding { family, .. } => {
+                assert_eq!(family, "Recoverable");
+            }
+        }
+    }
+
+    #[test]
+    fn early_return_template_returns_not_renderable() {
+        let result = NotNullPredicate.render(DropTemplate::EarlyReturn, "x");
+        let err = result.expect_err("EarlyReturn must return NotRenderable");
+        match err {
+            NotRenderable::Scaffolding { family, .. } => {
+                assert_eq!(family, "EarlyReturn");
+            }
+        }
+    }
+
+    #[test]
+    fn expect_template_returns_not_renderable() {
+        let result = NotNullPredicate.render(DropTemplate::Expect, "x");
+        let err = result.expect_err("Expect must return NotRenderable");
+        match err {
+            NotRenderable::Scaffolding { family, .. } => {
+                assert_eq!(family, "Expect");
+            }
+        }
+    }
+
+    #[test]
+    fn defensive_template_substitutes_var_name() {
+        let rendered = NotNullPredicate
+            .render(DropTemplate::Defensive, "my_var")
+            .expect("Defensive renders OK");
+        assert!(
+            rendered.contains("my_var"),
+            "Defensive template must contain var name 'my_var': {}",
+            rendered
+        );
+    }
+
+    #[test]
+    fn is_premise_guarded_true_for_is_some_premise() {
+        let wp = IrFormula::Implies {
+            operands: vec![
+                IrFormula::Atomic {
+                    name: "is_some".into(),
+                    args: vec![IrTerm::Var { name: "x".into() }],
+                },
+                IrFormula::Atomic {
+                    name: "not_null".into(),
+                    args: vec![IrTerm::Var { name: "x".into() }],
+                },
+            ],
+        };
+        assert!(NotNullPredicate.is_premise_guarded(&wp, "x"));
+    }
+
+    #[test]
+    fn is_premise_guarded_false_for_plain_not_null() {
+        let wp = IrFormula::Atomic {
+            name: "not_null".into(),
+            args: vec![IrTerm::Var { name: "x".into() }],
+        };
+        assert!(!NotNullPredicate.is_premise_guarded(&wp, "x"));
+    }
+
+    #[test]
+    fn is_premise_guarded_conservative_when_predicate_in_both() {
+        let wp = IrFormula::Implies {
+            operands: vec![
+                IrFormula::Atomic {
+                    name: "not_null".into(),
+                    args: vec![IrTerm::Var { name: "x".into() }],
+                },
+                IrFormula::Atomic {
+                    name: "not_null".into(),
+                    args: vec![IrTerm::Var { name: "x".into() }],
+                },
+            ],
+        };
+        assert!(!NotNullPredicate.is_premise_guarded(&wp, "x"));
+    }
+}
